+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/_template_builder.py
@@ -672,8 +672,8 @@ def build_vmss_resource(cmd, name, naming_prefix, location, tags, overprovision,
                         image=None, admin_password=None, ssh_key_value=None, ssh_key_path=None,
                         os_publisher=None, os_offer=None, os_sku=None, os_version=None,
                         backend_address_pool_id=None, inbound_nat_pool_id=None, health_probe=None,
-                        single_placement_group=None, custom_data=None, secrets=None, license_type=None,
-                        zones=None, priority=None):
+                        single_placement_group=None, platform_fault_domain_count=None, custom_data=None,
+                        secrets=None, license_type=None, zones=None, priority=None):
 
     # Build IP configuration
     ip_configuration = {
@@ -820,6 +820,10 @@ def build_vmss_resource(cmd, name, naming_prefix, location, tags, overprovision,
 
     if priority and cmd.supported_api_version(min_api='2017-12-01', operation_group='virtual_machine_scale_sets'):
         vmss_properties['virtualMachineProfile']['priority'] = priority
+
+    if platform_fault_domain_count is not None and cmd.supported_api_version(
+            min_api='2017-12-01', operation_group='virtual_machine_scale_sets'):
+        vmss_properties['platformFaultDomainCount'] = platform_fault_domain_count
 
     vmss = {
         'type': 'Microsoft.Compute/virtualMachineScaleSets',